% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availability.R
\name{angle_normalise}
\alias{angle_normalise}
\title{Fold angles into [-pi,pi)}
\usage{
angle_normalise(x)
}
\arguments{
\item{x}{a vector of angles}
}
\value{
a vector of folded angles
}
\description{
Fold angles (radians) into [-pi,pi)
}
\details{
This is an internal function used by
\code{\link{surrogate_arsimulate}}.
}
\author{
Ben Raymond
}

