% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newparm.R
\name{av_fit_ar}
\alias{av_fit_ar}
\title{Fit first-order vector-autoregressive model to track}
\usage{
av_fit_ar(x, ...)
}
\arguments{
\item{x}{: a data.frame containing the observed track. It can optionally be an \code{sf}-tibble (i.e. in projected coordinates). If not projected, it must contain columns named "lon" and "lat". In both cases it must contain a "date" column with timestamps in POSIX format}

\item{...}{: additional arguments, currently not documented}
}
\value{
An object of class "av_fit"
}
\description{
Fit first-order vector-autoregressive model to track
}
\seealso{
\code{\link[=surrogateAR]{surrogateAR()}} for the original (differently-parameterized) version of this function, and \code{\link[=ar]{ar()}} for the underlying function that fits the VAR model to the step speeds
}
