% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability-package.R
\docType{package}
\name{availability}
\alias{availability-package}
\alias{availability}
\title{\pkg{availability}}
\description{
Estimating geographic space available to animals based on telemetry data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AustralianAntarcticDivision/availability}
  \item Report bugs at \url{https://github.com/AustralianAntarcticDivision/availability/issues}
}

}
\author{
\strong{Maintainer}: Ben Raymond \email{ben.raymond@aad.gov.au}

Authors:
\itemize{
  \item Ben Raymond \email{ben.raymond@aad.gov.au}
  \item Simon Wotherspoon
  \item Ian Jonsen
  \item Ryan Reisinger
}

}
