% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availability.R
\name{calc_distbearing}
\alias{calc_distbearing}
\title{Calculate distances in metres and bearing between successive points}
\usage{
calc_distbearing(lonlat)
}
\arguments{
\item{lonlat}{array: 2-column matrix or data.frame with longitude, latitude of each point}
}
\value{
data.frame with distance in metres and bearing in degrees
}
\description{
Calculate distances in metres and bearing between successive points
}
\author{
Ben Raymond
}

