% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availability.R
\name{calc_dxdy}
\alias{calc_dxdy}
\title{Calculate x- and y-step distances in metres between successive points}
\usage{
calc_dxdy(lonlat)
}
\arguments{
\item{lonlat}{array: 2-column matrix or data.frame with longitude, latitude of each point}
}
\value{
data.frame with dx and dy in metres
}
\description{
Calculate x- and y-step distances in metres between successive points
}
\author{
Ben Raymond
}

