% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landmask.R
\name{gebcoMask}
\alias{gebcoMask}
\title{Land Mask}
\usage{
gebcoMask(path = "c:/Gebco/", land = FALSE)
}
\arguments{
\item{path}{the path to a folder containing gebco_08.tif}

\item{land}{logical - the value to return for land.}
}
\value{
a logical indicating whether the point is land or sea.
}
\description{
A land mask based on Gebco 08.
}
\details{
Generate a land mask function based on Gebco 08 topography.  The
mask is constant, the \code{tm} argument to the mask is
ignored. The \code{land} argument determines whether the mask
function returns \code{TRUE} or \code{FALSE} for land.
}

