% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landmask.R
\name{landmask_init}
\alias{landmask_init}
\title{Initialize the land-masking function}
\usage{
landmask_init()
}
\value{
function
}
\description{
Initialize the land-masking function
}
\examples{
lm=landmask_init() ## initialize land mask function
lm(c(100,-65)) ## test point lon,lat
}
\seealso{
\code{\link{surrogate_arsimulate}}
}

