% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ptt_data}
\alias{ptt_data}
\title{Example PTT tracking data}
\format{
A list containing three tracks collected using Argos PTT tags: \code{DMS} from a sooty shearwater, \code{SES} from an elephant seal, and \code{AFS} from an Antarctic fur seal.
}
\source{
\href{https://doi.org/10.1111/ddi.12702}{Reisinger RR, et al. (2018) Habitat modelling of tracking data from multiple marine top predators reveals important habitat in the Southern Indian Ocean. Diversity and Distributions.}
}
\usage{
data(ptt_data)
}
\description{
Three example animal tracks, from Reisinger et al. (2018)
}
\references{
https://doi.org/10.1111/ddi.12702
}
\keyword{datasets}
