% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{randomize_track}
\alias{randomize_track}
\title{Randomized track}
\usage{
randomize_track(lonlat, rotate = c(-pi, pi), reorder = FALSE)
}
\arguments{
\item{lonlat}{a 2-column matrix or dataframe with longitude and latitude of each point}

\item{rotate}{a 2-element numeric vector giving the lower and upper limits of the random rotation to apply to the randomized track}

\item{reorder}{should the track steps be randomly reordered?}
}
\value{
A dataframe with columns:
\itemize{
\item \code{lon}: the longitude of the randomized track
\item \code{lat}: the latitude of the randomized track
}
}
\description{
Compute surrogate track by randomizing the steps of an observed track
}
\details{
Converts a track to a distance/bearing representation, and then
reconstructs a new track by randomly perturbing the bearings of
each increment, and otpionally, randomly reordering the
increments.
}
