% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availability.R
\name{surrogate_arfit}
\alias{surrogate_arfit}
\title{VAR(1) track model}
\usage{
surrogate_arfit(lonlat)
}
\arguments{
\item{lonlat}{a 2-column matrix or dataframe with longitude and
latitude of each point}
}
\value{
An object of class "ar"
}
\description{
Fit first-order vector-autoregressive model to track
}
\details{
This function fits the vector AR(1) model used to model the track
incements in \code{\link{surrogate_arsimulate}}.
}
\author{
Ben Raymond
}
\seealso{
\code{\link{ar}}, \code{\link{surrogate_arsimulate}}
}

