% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animotum.R
\name{surrogateAM}
\alias{surrogateAM}
\title{aniMotum bridge sampler}
\usage{
surrogateAM(
  fit,
  xs,
  ts,
  fixed = rep(c(TRUE, FALSE, TRUE), times = c(1, nrow(xs) - 2, 1)),
  point.check = function(tm, pt) TRUE,
  partial = FALSE
)
}
\arguments{
\item{fit}{a fitted model generated with \code{\link[aniMotum:fit_ssm]{aniMotum::fit_ssm()}}}

\item{xs}{matrix: a two-column matrix giving the the template sequence of locations (longitude and latitude). If not provided, the "predicted" (i.e. filtered and interpolated) track from the \code{fit} object will be used}

\item{ts}{POSIXct: the times at which the track is sampled}

\item{fixed}{logical: a vector (with length equal to the number of rows in \code{xs}) indicating which locations in the template path are to be held fixed}

\item{point.check}{function: a function that accepts a state and returns \code{TRUE} or \code{FALSE} indicating whether the state is acceptable}

\item{partial}{if \code{TRUE}, a partial track is returned if the sampling fails}
}
\value{
An array of states the define the simulated path
}
\description{
Generate new tracks from a fitted aniMotum model
}
\details{
Given a fitted aniMotum crw or rw model, and optionally a template track, this function generates a new track of the same length that coincides with the template at the start point and optionally other specified points along the track. This function works similarly to \code{\link[aniMotum:sim_fit]{aniMotum::sim_fit()}} but with different handling of land masking and fixed points.

Additional constraints can be placed on the path by rejection sampling through the function \code{point.check}.  This function must accept a state and return a boolean indicating whether the point is acceptable.  For example, the track can be constrained to the ocean by supplying a \code{point.check} function that compares the state to a land mask and returns \code{FALSE} for states corresponding to locations that fall on land.
}
