% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{surrogateARModel}
\alias{surrogateARModel}
\title{VAR(1) track model}
\usage{
surrogateARModel(lonlat)
}
\arguments{
\item{lonlat}{a 2-column matrix or data.frame with longitude and latitude of each point}
}
\value{
An object of class "ar"
}
\description{
Fit first-order vector-autoregressive model to track
}
\details{
This function fits the vector AR(1) model used as the \code{model} argument to \code{\link[=surrogateAR]{surrogateAR()}}
}
\seealso{
\code{\link[=ar]{ar()}}, \code{\link[=surrogateAR]{surrogateAR()}}
}
