% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{surrogateCrawlModel}
\alias{surrogateCrawlModel}
\title{Crawl Movement Model}
\usage{
surrogateCrawlModel(fit, dt)
}
\arguments{
\item{fit}{a fitted crawl object}

\item{dt}{the fixed time increment}
}
\value{
A list containing
\item{\code{A}}{the transition matrix}
\item{\code{Q}}{the covariance matrix}
\item{\code{dt}}{the time increment}
}
\description{
Generate Transition and Covariance Matrices for a simple Crawl Model
}
\details{
Generates the transition and covariance matrices of the state
space model for time increment \code{dt}, corresponding to a
fitted crawl model with no drift and no covariates.

Currently no polar adjustment is made.

The matrices are generated assuming the state variables are stored
as a vector in the order longitude mu and nu then latitude mu and
nu.  The parametrization is defined as in the crawl source not the
paper, which differs by a factor of \eqn{\beta^{2}} in the
definition of \eqn{\sigma^{2}}.
}
