% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability-package.R
\docType{package}
\name{availability}
\alias{availability}
\title{\pkg{availability}}
\description{
Estimating geographic space available to animals based on telemetry data
}
\author{
Ben Raymond \email{ben.raymond@aad.gov.au}, Simon Wotherspoon, Ryan Reisinger
}
