% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{calc_distbearing}
\alias{calc_distbearing}
\title{Distance and Bearing}
\usage{
calc_distbearing(lonlat)
}
\arguments{
\item{lonlat}{a 2-column matrix or dataframe with longitude and
latitude of each point}
}
\value{
A dataframe with columns
\item{\code{lon}}{the longitude of the randomized track}
\item{\code{lat}}{the latitude of the randomized track}
}
\description{
Calculate distances in metres and bearing between successive
points along track.
}
\details{
This is an internal function used by \code{\link{randomize_track}}.
}
\author{
Ben Raymond
}

