% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmask.R
\name{etopoMask}
\alias{etopoMask}
\title{Land Mask}
\usage{
etopoMask(basename = "ETOPO1_Bed_c_geotiff", path = ".", tmp = path,
  land = FALSE)
}
\arguments{
\item{basename}{the name of the etopo geotiff (without file extension).}

\item{path}{the path to a folder containing the etopo geotiff}

\item{tmp}{the path to a writeable folder}

\item{land}{the logical value to return for land.}
}
\value{
a logical indicating whether the point is land or sea.
}
\description{
A land mask based on the full ETOPO1 dataset
}
\details{
Generate a land mask function based on ETOPO1 topography. The
etopo geotiff is not bundled with the package and must be
downloaded from
\url{https://www.ngdc.noaa.gov/mgg/global/global.html}.

When the mask is intially created, a native raster (grd,gri)
version of the geotiff is created in the directory \code{tmp},
which must be writable. This file can be deleted when the
computation is finished.

The \code{land} argument determines whether the mask function
returns \code{TRUE} or \code{FALSE} for land. The mask is constant
and the \code{tm} argument to the mask is ignored.
}
