% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmask.R
\name{gshhsMask}
\alias{gshhsMask}
\alias{landmask_init}
\title{A land mask based on the GSHHS data set}
\usage{
gshhsMask(res = 0.1)

landmask_init(res = 0.1)
}
\arguments{
\item{res}{numeric: the spatial resolution of the mask, in degrees (either 0.1 or 0.05)}
}
\value{
function that returns a logical indicating whether the point is at sea (TRUE) or on land (FALSE)
}
\description{
Generate a land mask function based on the Global Self-consistent, Hierarchical, High-resolution Geography Database.
The mask is provided at two (approximate) spatial resolutions: 0.1 degree and 0.05 degrees. The latter requires significantly more memory.
The mask is constant and the \code{tm} argument to the mask is ignored.
}
\examples{
mask  <- gshhsMask() ## initialize land mask function
mask(0, c(100, -65)) ## test point lon,lat

}
\references{
Wessel P, Smith WHF (1996) A Global Self-consistent, Hierarchical, High-resolution Shoreline Database. J. Geophys. Res. 101: 8741-8743. \url{https://www.ngdc.noaa.gov/mgg/shorelines/gshhs.html}
}
\seealso{
\code{\link{surrogateAR}}
}
