% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availability.R
\name{randomize_track}
\alias{randomize_track}
\title{Compute surrogate track by randomizing the steps of an observed track}
\usage{
randomize_track(lonlat, rotate = c(-pi, pi))
}
\arguments{
\item{lonlat}{array: 2-column matrix or data.frame with longitude, latitude of each point}

\item{rotate}{array: 2-element numeric giving the lower and upper limits of the random rotation to apply to the randomized track}
}
\value{
2-column data.frame with longitude,latitude of randomized track points
}
\description{
Assumes that steps are equally spaced in time. Probably not a particular good way of doing it since it destroys the autocorrelation
structure of the track.
}
\author{
Ben Raymond
}

