% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availability.R
\name{surrogate_arfit}
\alias{surrogate_arfit}
\title{Fit vector-autoregressive model to track}
\usage{
surrogate_arfit(lonlat, model.order = 1)
}
\arguments{
\item{lonlat}{array: 2-column matrix or data.frame with longitude, latitude of each point}

\item{model.order}{numeric: the order of the AR model to fit, default=1}
}
\value{
object of class "ar"
}
\description{
Fit vector-autoregressive model to track
}
\author{
Ben Raymond
}
\seealso{
\code{\link{ar}} \code{\link{surrogate_arsimulate}}
}

